+++ lab_7/lab_7.c
@@ -16,7 +16,7 @@ void* readT(){
 	for(; ; ){
 		pthread_mutex_lock(&mutex);
 		pthread_cond_wait(&cond, &mutex);
-		printf("ID потока: %ld, Размер массива: %d, Счетчик: %d\n", pthread_self(), arr[counter], counter);
+		printf("ID потока: %ld, Размер массива: %d, Счетчик: %d\n", (long)pthread_self(), arr[counter], counter);
 		
 		fflush(stdout);
